final styles for task form
